rename sensors to instruments
